Fix that starts_at could be nil for live articles.
If an article has its status set to PUBLISHED manually, it might not
have a starts_at date. This can be used throughout the site as the
publication date, without nil checks (as on The Skinny).

This fix validates that starts_at is not nil for published articles, and
adds a before_validation callback to set it to a non-nil value if
necessary.